+++ src/mako/database/query/SortDirection.php
@@ -1,0 +1,17 @@
+<?php
+
+/**
+ * @copyright Frederic G. Østby
+ * @license   http://www.makoframework.com/license
+ */
+
+namespace mako\database\query;
+
+/**
+ * Sort direction.
+ */
+enum SortDirection: string
+{
+	case Ascending = 'ASC';
+	case Descending = 'DESC';
+}