+++ src/mako/database/query/compilers/SQLite.php
@@ -45,6 +45,17 @@ class SQLite extends Compiler
 	 * {@inheritDoc}
 	 */
 	#[Override]
+	protected function nullOrdering(array $order): string
+	{
+		$nullOrder = $order['nullsLast'] ? 'LAST' : 'FIRST';
+
+		return "{$this->column($order['column'])} {$order['order']} NULLS {$nullOrder}";
+	}
+
+	/**
+	 * {@inheritDoc}
+	 */
+	#[Override]
 	protected function betweenDate(array $where): string
 	{
 		$date1 = "{$where['value1']} 00:00:00.000";